Refactor: Split cli/__init__.py into command modules
Split the 2093-line cli/__init__.py into a modular package structure:

- __init__.py (302 lines): Main dispatch and entry points
- parser.py (1071 lines): All argparse configuration
- utils.py (28 lines): Shared utilities (format_error)
- commands/ package: 11 modules with domain-specific handlers
  - schematic.py: sch subcommand handlers
  - pcb.py: pcb subcommand handlers
  - library.py: lib subcommand handlers
  - routing.py: route, zones, optimize-traces handlers
  - validation.py: check, validate, footprint validation
  - footprint.py: footprint generation
  - parts.py: LCSC parts lookup
  - datasheet.py: datasheet commands
  - reasoning.py: reason command
  - config.py: config and interactive commands
  - manufacturer.py: mfr subcommand handlers
  - placement.py: placement commands

Benefits:
- No single module exceeds 1100 lines (was 2093)
- Command handlers are 30-140 lines each
- Easier to add new commands
- Better testability and maintainability
- Public API (main(), entry points) unchanged

All 49 CLI tests pass.
